+++ unirest/src/test/java/BehaviorTests/MockServer.java
@@ -63,7 +63,8 @@ public class MockServer {
 	public static final String GZIP = HOST + "/gzip";
 	public static final String PATCH = HOST + "/patch";
 	public static final String INVALID_REQUEST = HOST + "/invalid";
-	public static final String PASSED_PATH_PARAM = GET + "/{param}/passed";
+	public static final String PASSED_PATH_PARAM = GET + "/{params}/passed";
+	public static final String PASSED_PATH_PARAM_MULTI = PASSED_PATH_PARAM + "/{another}";
 	public static final String CHEESE = HOST + "/cheese";
 	public static final String ALTGET = "http://127.0.0.1:" + PORT + "/get";
 
@@ -94,7 +95,8 @@ public class MockServer {
 		get("/nobody", MockServer::nobody);
 		head("/get", MockServer::jsonResponse);
 		put("/post", MockServer::jsonResponse);
-		get("/get/:p/passed", MockServer::jsonResponse);
+		get("/get/:params/passed", MockServer::jsonResponse);
+		get("/get/:params/passed/:another", MockServer::jsonResponse);
 		get("/proxy", MockServer::proxiedResponse);
 		get("/binary", MockServer::file);
 		get("/paged", MockServer::paged);